Move calendarUtils/getCalendarEnd to calendar package
