+++ matrix-meetings-widget/src/lib/utils/calendarUtils/index.ts
@@ -17,7 +17,6 @@
 export { calculateCalendarEvents } from './calculateCalendarEvents';
 export type { CalendarSourceEntries } from './calculateCalendarEvents';
 export { deleteCalendarEvent } from './deleteCalendarEvent';
-export { getCalendarEnd } from './getCalendarEnd';
 export { getCalendarEvent } from './getCalendarEvent';
 export {
   isRecurringCalendarSourceEntry,
