+++ packages/calendar/package.json
@@ -20,10 +20,12 @@
   },
   "dependencies": {
     "joi": "^17.11.0",
+    "lodash": "^4.17.20",
     "luxon": "^3.3.0",
     "rrule": "^2.8.1"
   },
   "devDependencies": {
+    "@types/lodash": "^4.14.200",
     "@types/luxon": "^3.2.0"
   },
   "jest": {
